week05/02-MusicPlayer/requirements.txt
click==6.7
mutagen==1.40.0
python-mpd2==1.0.0
tabulate==0.8.2
